Move Windows-specific config settings into compat/mingw.c
